Open slides in browser.

Call this function with the session number as argument and the slides of that session will open in your browser (an internet connection is required).

open_slides(session)

Arguments

session
Session number as a number.

Value

Nothing. Used for its side effects.

Examples

if (FALSE) {
open_slides(0)
}