The phonetic profile of Korean formal and informal speech registers.

This tibble contains data from the study discussed in Winter and Grawunder 2012.

polite

Format

A tibble with 224 rows and 29 variables:

subject

Subject unique identifier [categorical].

gender

Gender of subject [categorical].

birthplace

Birth place of subject [categorical].

musicstudent

Does the subject have music training? [binary: yes, no]

scenario

Unique identifier of different items.

task

Task type [categorical: not = mailbox task vs dct = discourse completion task]. In the mailbox task people left a note on somebody's mailbox, while in the discourse completion task they were prompted to role-play the start of a conversation.

attitude

Attitude [binary: polite vs informal].

total_duration

Total duration of utterances in seconds [numeric].

articulation_rate

Number of syllables per second [numeric].

f0mn

Mean fundamental frequency (f0) [numeric].

f0sd

Stadard deviation of fundamental frequency [numeric].

f0range

Minimum and maximum fundamental frequency [numeric].

inmn

Mean intensity [numeric].

insd

Standard deviation of intensity [numeric].

inrange

Minimum and maximum fundamental frequency [numeric].

shimmer

Local shimmer (likewise normalized amplitude difference of consecutive periods) [numeric].

jitter

Local jitter (bsolute period-to-period difference divided by the average period) [numeric].

HNRmn

Mean Harmonics-to-Noise Ratio [numeric].

H1H2

Difference between first and second harmonic (H1-H2) [numeric].

breath_count

Number of audible breath intakes [count].

filler_count

Number of oral fillers like "oh/ah" [count].

hiss_count

Number of noisy breath intakes [count].

nasal_count

Number of nasal fillers like "mh/nh" [count].

sil_count

Number of silent pauses [count].

ya_count

Number of occurences of interjection "ya" (informal) [count].

yey_count

Number of occurences of interjection "yey" (polite) [count].

Source

Bodo Winter, Sven Grawunder. 2012. The phonetic profile of Korean formal and informal speech registers, Journal of Phonetics 40(6). 808-815. https://doi.org/10.1016/j.wocn.2012.08.006